script 0

+++ file_io/0-read_textfile.c
@@ -9,13 +9,13 @@
 ssize_t read_textfile(const char *filename, size_t letters)
 {
 	FILE *ptr;
-	int count = 0;
+	size_t count = 0;
 	char ch;
 
 	if (filename)
 		ptr = fopen(filename, "r");
 
-	while (ptr && ch != EOF && count < (int)letters)
+	while (ptr && ch != EOF && count < letters)
 	{
 		ch = fgetc(ptr);
 		_putchar(ch);
